Introduce record combining position and range

+++ src/main/java/com/mraof/minestuck/computer/editmode/ClientEditHandler.java
@@ -146,7 +146,7 @@ public final class ClientEditHandler
 		{
 			Player player = event.player;
 			
-			double range = ClientDimensionData.isLand(player.level().dimension()) ? MinestuckConfig.SERVER.landEditRange.get() : MinestuckConfig.SERVER.overworldEditRange.get();
+			int range = ClientDimensionData.isLand(player.level().dimension()) ? MinestuckConfig.SERVER.landEditRange.get() : MinestuckConfig.SERVER.overworldEditRange.get();
 			
 			if(locations != null)
 				locations.limitMovement(player, range);
